requirements.txt
altgraph==0.17.4
MouseInfo==0.1.3
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32-ctypes==0.2.3
six==1.16.0
